apps/analytics/requirements.txt
pandas
numpy
azure-storage-blob
azure-identity
pytest
